:recycle: (sonarcloud): Fix Global variables should be const
Motors were created in the global scope in the MotorsUtils header which
was not ideal nor good programming...

This commit moves motors to main() and refactors the motors related
functions by passing a Motors struct as a pointer.

This also allows the use of Callback to pass the function to thread.