:rewind: (workflows): Revert running on pull_request_target
The assumption was that Github Actions would decide to use
pull_request_target only for PR from forks but it doesn't work this way.
Instead it just duplicated the workflows like a matrice which is not
what we wanted to do.

This seems to cause issues with Sonarcloud and other tools as the report
are uploaded twice with different parameters.

The solution would be to duplicate the workflows, one for pull_request
and one for pull_request_target but it adds a lot of maintainance costs.
Colaborators will just work on the main repo for the time being.