:fire: (tests): Remove functional tests
Those tests were used for the certification process and to test some
hardware parts of the robot (i.e. motors)

They were old, not maintained and not up-to-date with our coding
standards.

They have been repalced and/or superseded by spikes testing all the
functionalities of the robot.

They could be reintroduced if needed in the future, but for now, it is
better to clean them up to avoid a huge technical debt.