:sparkles: (CommandKit): Use EventLoopKit to execute command in different context
fixes #640

CommandKit now executes the commands in a different context provided by
EventLoopKit. This avoids deadlocks, ISR locks and other (BLE) bugs that
we encountered.

The event queue was tested but the result was not satisfying: as we
where using pointers, the command was changing between the call and the
actual execution of the command.