:adhesive_bandage: (VideoKit): Use mutex to prevent race conditions
On image + video display, we are experiencing a strange bug due to a race
condtion which results in the robot being stuck with a non working LCD screen
and the code running forever in a while(true) loop in CoreJPEG::findSOIMarker.

This is due to `_image_properties` not being updated when tryin to
display an image and a video at the same time.

Using a mutex to lock the access to the shared resource fixes the issue