:recycle: (VideoKit): Move mutex to anonymous namespace in .cpp
This gives the mutex static storage duration, and makes it available to
the current TU only, in our case VideoKit.cpp

This will makes sure that if multiple instances of VideoKit exist, the
mutex will be shared by all of them preventing race conditions.

Another option would be to use the mutex as a private static member
variable but this has the downside of polluting the public interface and
header files, and exposing implementation details.