:adhesive_bandage: (power mgmt): Disable input for LogKit's default_serial
SerialBuffer, by default, holds a lock on deepsleep preventing the system
from deepsleeping to wait for incoming data, thus preventing the system
from saving powser when needed.

In our case, nothing is ever sent to the robot through UART, so input
can be safely disabled.

This has the effect to allow deep sleep when using LogKit.

If at some point we would like to send data through UART, it can be
enabled back on when needed.

See link for more information:

https://os.mbed.com/docs/mbed-os/v6.15/apis/power-management-sleep.html