:zap: (power): LogKit - Disable default_serial stdio input
SerialBuffer, by default, holds a lock on deep sleep preventing the system
from deep sleeping to wait for incoming data, thus preventing the system
from saving power when needed.

In our case, nothing is ever sent to the robot through UART, so input
can be safely disabled.

This has the effect to allow deep sleep when using LogKit.

If at some point we would like to send data through UART, it can be
enabled back on when needed.

See link for more information:

https://os.mbed.com/docs/mbed-os/v6.15/apis/power-management-sleep.html