:wrench: (BLE): Cordio - set max-prepared-writes to 1
This setting determines how many prepare write requests can be queued before they are executed.
Each prepare write request can contain a segment of the data to be written to an attribute.

When cordio.max-prepared-writes is set to 1, it means your system can only hold one prepare
write request at a time. This setting minimizes memory usage but can slow down the process
if the device needs to wait for each segment to be confirmed before sending the next.

Default value is 4

With the default value, we were often receiving 4 packets at the same
time. For DFU, this meant adding more logic to handle the 4 packets
correctly which we deemed too much work for now.

Setting the value to 1 prevents mutliple packets from being received at
once and breaking the DFU process.